Fix issue where API call was being made repeatedly
The way I was using React Router was causing a GET request to be made to the server for category and item detail over and over upon loading the page.

These changes fix that. In addition, I figured out how to make API calls before anything renders on the page for functional components by short-circuiting the component (in the absence of a componentWillMount lifecycle method or a class-based component's constructor function).